+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## 3.2.0 - 2023-02-11
+
+### Added
+
+- `Innmind\TimeContinuum\Earth\Period`
+- `Innmind\TimeContinuum\Earth\ElapsedPeriod::ofPeriod()`
+- `Innmind\TimeContinuum\Earth\Period\Millisecond::asElapsedPeriod()`
+- `Innmind\TimeContinuum\Earth\Period\Second::asElapsedPeriod()`
+- `Innmind\TimeContinuum\Earth\Period\Minute::asElapsedPeriod()`
+- `Innmind\TimeContinuum\Earth\Period\Hour::asElapsedPeriod()`
+- `Innmind\TimeContinuum\Earth\Period\Day::asElapsedPeriod()`
+- `Innmind\TimeContinuum\Earth\Period\Composite::asElapsedPeriod()`
+
+### Changed
+
+- Require PHP `8.1`
+
 ## 3.1.0 - 2022-11-14
 
 ### Added